refactor: move base integration test to separate package
