+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/GrantAccessIntegrationTest.java
@@ -1,5 +1,6 @@
 package integration.uk.gov.hmcts.reform.amlib;
 
+import integration.uk.gov.hmcts.reform.amlib.base.IntegrationBaseTest;
 import org.junit.Before;
 import org.junit.Test;
